refactor:#7 userId 가져오는 로직 분리
