--- src/components/commons/NaverMap.jsx
@@ -1,5 +1,0 @@
-const NaverMap = () => {
-  return <div id='map' className="w-full h-screen"></div>;
-};
-
-export default NaverMap;
