+++ src/components/features/create-plan-page/NaverMap.jsx
@@ -1,0 +1,7 @@
+import { planMapFile } from "../../../constants/naverMap";
+
+const NaverMap = () => {
+  return <div id={planMapFile.MAPID} className="w-full h-screen"></div>;
+};
+
+export default NaverMap;
