+++ src/constants/naverMap.js
@@ -12,7 +12,8 @@ export const planMapFile = {
 };
 
 export const markerFile = {
-  ANCHORPOINT: 15,
   ZOOM: 3,
-  CLICK: 'click'
+  CLICK: 'click',
+  ANCHORPOINT_X: 15,
+  ANCHORPOINT_Y: 40
 };
