+++ src/lib/hooks/create-plan/useMapClickListner.js
@@ -1,8 +1,9 @@
 import { useEffect } from 'react';
+import { markerFile } from '../../../constants/naverMap';
 
 export const useMapClickListner = (mapRef, setMarker) => {
   useEffect(() => {
-    window.naver.maps.Event.addListener(mapRef.current, 'click', (e) => {
+    window.naver.maps.Event.addListener(mapRef.current, markerFile.CLICK, (e) => {
       setMarker({
         position: e.coord,
         number: null
